tests/wpt/web-platform-tests/navigation-api/updateCurrentEntry-method/initial-about-blank.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe"></iframe>

<script>
test(() => {
  assert_throws_dom("InvalidStateError", iframe.contentWindow.DOMException, () => {
    iframe.contentWindow.navigation.updateCurrentEntry({ state: 1 });
  });
  assert_equals(navigation.currentEntry.getState(), undefined);
}, "updateCurrentEntry() must throw if the document is still on the initial about:blank");
</script>
