LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/initial-about-blank-expected.txt



PASS updateCurrentEntry() must throw if the document is still on the initial about:blank
